search screen style optimization
