+++ src/screens/SearchScreen.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
-import { ActivityIndicator, Platform, View, StyleSheet, Text, FlatList } from 'react-native';
+import { ActivityIndicator, Platform, View, StyleSheet, Text, FlatList, Dimensions } from 'react-native';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
+import { Loading } from '../components/Loading';
 import { PokemonCard } from '../components/PokemonCard';
 import { SearchInput } from '../components/SearchInput';
 import { usePokemonSearch } from '../hooks/usePokemonSearch';
@@ -8,31 +9,32 @@ import { style } from '../theme/appTheme';
 
 export const SearchScreen = () => {
 
+    const screenWidth = Dimensions.get('window').width;
+
     const { top } = useSafeAreaInsets();
 
     const { simplePokemonList, isFetching } = usePokemonSearch();
 
     if(isFetching){
-        return (
-            <View style={ styleSearch.activityContainer }>
-                <ActivityIndicator 
-                    size={ 40 }
-                    color='#808080'
-                />
-                <Text style={{ marginTop: 10, alignItems: 'center' }}>Loading...</Text>
-            </View>
-        )
+        return <Loading />
     }
 
     return (
         <View 
             style={{ 
                 flex: 1, 
-                marginTop: (Platform.OS === 'ios') ? top : top + 10
+                // marginTop: (Platform.OS === 'ios') ? top : top + 10
             }}
         >
             <View style={{ marginHorizontal: 20 }}>
-                <SearchInput />
+                <SearchInput 
+                    style={{
+                        position: 'absolute',
+                        zIndex: 999,
+                        width: screenWidth - 40,
+                        marginTop: (Platform.OS === 'ios') ? top : top + 20
+                    }}
+                />
             </View>
 
             <View style={{ alignItems: 'center' }}>
@@ -47,7 +49,8 @@ export const SearchScreen = () => {
                             style={{
                                 ...style.title,
                                 ...style.globalMargin,
-                                paddingBottom: 10
+                                paddingBottom: 10,
+                                marginTop: (Platform.OS === 'ios') ? top + 65 : top + 65
                             }}
                         >
                             PokeDex
@@ -63,11 +66,3 @@ export const SearchScreen = () => {
         </View>
     )
 };
-
-const styleSearch = StyleSheet.create({
-    activityContainer: {
-        flex: 1,
-        justifyContent: 'center',
-        alignItems: 'center'
-    }
-});
